{{cookiecutter.project_slug}}/requirements_dev.txt
pip==19.2.3
bump2version==0.5.11
wheel==0.33.6
watchdog==0.9.0
flake8==3.7.8
tox==3.14.0
coverage==4.5.4
Sphinx==1.8.5
twine==1.14.0
pytest==5.3.1
cookiecutter>=1.4.0
pytest-cookies==0.5.1
alabaster==0.7.12
pydantic>=2.0.0
pyyaml>=6.0.0
{% if cookiecutter.command_line_interface|lower == 'click' -%}
Click==7.1.2{% endif %}
{% if cookiecutter.use_pytest == 'y' -%}
pytest==6.2.4{% endif %}
{% if cookiecutter.use_black == 'y' -%}
black==21.7b0{% endif %}
